+++ limiter/sharedquota_test.go
@@ -7,7 +7,7 @@ import (
 	"testing"
 	"time"
 
-	"github.com/jaeminst/pace/internal/breaker"
+	"github.com/jaeminst/pace/breaker"
 	pace "github.com/jaeminst/pace/limiter"
 	"github.com/jaeminst/pace/observe"
 	"github.com/jaeminst/pace/rate"